requirements.txt
argcomplete==1.12.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
commitizen==2.20.4
decli==0.5.2
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-aiplatform==1.9.0
google-cloud-bigquery==2.32.0
google-cloud-core==2.2.2
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.1
opencv-python==4.5.5.62
packaging==21.3
prompt-toolkit==3.0.26
proto-plus==1.19.9
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==6.0
questionary==1.10.0
requests==2.27.1
rsa==4.8
six==1.16.0
termcolor==1.1.0
tomlkit==0.8.0
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
